don't send empty list of nodes when restricting network traffic
When sending empty list to method 'set_allowed_peers' we allow for all connections.
Problem appears when there is only one node in network and we want to restrict network
traffic for this network (disallow all connections). Previously we would send empty list in this case,
effectively doing nothing. With this fix we send node self id, effectively disallowing for all connections.