Remove checking if wallet responds to messages at startup
In the past there was a problem with wallet, that when just after start user
sent message to wallet, then wallet sometimes responded with error. It caused
problems with writing tests, because wallet behavior was unstable.

Lastly I tried to remove that check and turns out it is not needed anymore.
I tested it hundreds of times with different builds and I never reproduced the
problem.