Rename String to StringQuoted and Untouched to String
To express "string which will be written to config as provided" user have to
write `String()`. E.g. string 'abc' will be serialized in config to:
  example_config_entry = abc

To express "string with double quotes as first and last characters" user have to
write `StringQuoted()`. E.g. string 'abc' will be serialized in config to:
  example_config_entry = "abc"

In both cases user in Python code always gets and sets strings without any
additional quotation marks. So user writes always
  `node.config.example = 'abc'`
and it is NodeConfig's responsibility to serialize it properly.