Prevent copying block log files partially when error occurs
There was a bug, that first block_log file was copied, and then existing of
artifacts file was checked. If artifacts were missing, then error was raised,
but part of block log was already copied.

Now copying procedure ensures, that files to copy exists and copying will
pass without problems.